Trying out JXTable
