+++ gradle/libs.versions.toml
@@ -2,14 +2,8 @@
 java = "17"
 kotlin = "1.6.10"
 coroutines = "1.6.0"
-serialization = "1.3.1"
-miglayout = "11.0"
-flatlaf = "1.6.2"
-svgSalamander = "1.1.2.4"
-jide = "3.6.18"
-logback = "1.2.3"
+flatlaf = "2.0.1"
 kotest = "5.0.3"
-zip4j = "2.9.1"
 ignition = "8.1.1"
 
 [plugins]
@@ -22,19 +16,21 @@ runtime = { id = "org.beryx.runtime", version = "1.12.7" }
 # core functionality
 coroutines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "coroutines" }
 coroutines-swing = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-swing", version.ref = "coroutines" }
-serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "serialization" }
+serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version = "1.3.1" }
 xerial-jdbc = { group = "org.xerial", name = "sqlite-jdbc", version = "3.36.0.3" }
-logback = { group = "ch.qos.logback", name = "logback-classic", version.ref = "logback" }
+logback = { group = "ch.qos.logback", name = "logback-classic", version = "1.2.3" }
 hsql = { group = "org.hsqldb", name = "hsqldb", version = "2.6.1" }
-zip4j = { group = "net.lingala.zip4j", name = "zip4j", version.ref = "zip4j" }
+zip4j = { group = "net.lingala.zip4j", name = "zip4j", version = "2.9.1" }
 
 # appearance/swing
-miglayout = { group = "com.miglayout", name = "miglayout-swing", version.ref = "miglayout" }
+miglayout = { group = "com.miglayout", name = "miglayout-swing", version = "11.0" }
 flatlaf-core = { group = "com.formdev", name = "flatlaf", version.ref = "flatlaf" }
 flatlaf-extras = { group = "com.formdev", name = "flatlaf-extras", version.ref = "flatlaf" }
 flatlaf-jide = { group = "com.formdev", name = "flatlaf-jide-oss", version.ref = "flatlaf" }
-svgSalamander = { group = "com.formdev", name = "svgSalamander", version.ref = "svgSalamander" }
-jide-common = { group = "com.jidesoft", name = "jide-oss", version.ref = "jide" }
+flatlaf-swingx = { group = "com.formdev", name = "flatlaf-swingx", version.ref = "flatlaf" }
+svgSalamander = { group = "com.formdev", name = "svgSalamander", version = "1.1.2.4" }
+jide-common = { group = "com.jidesoft", name = "jide-oss", version = "3.6.18" }
+swingx = { group = "org.swinglabs.swingx", name = "swingx-core", version = "1.6.5-1" }
 
 # Ignition
 ignition-common = { group = "com.inductiveautomation.ignition", name = "common", version.ref = "ignition" }
@@ -58,6 +54,7 @@ flatlaf = [
     "flatlaf-core",
     "flatlaf-extras",
     "flatlaf-jide",
+    "flatlaf-swingx",
 ]
 kotest = [
     "kotest-junit",
