--- src/main/kotlin/io/github/paulgriffith/idb/logviewer/LogExportTable.kt
@@ -1,39 +1,0 @@
-package io.github.paulgriffith.idb.logviewer
-
-import io.github.paulgriffith.idb.logviewer.LogExportModel.EventColumns.Timestamp
-import io.github.paulgriffith.utils.setDefaultRenderer
-import io.github.paulgriffith.utils.setupColumns
-import javax.swing.JTable
-import javax.swing.RowSorter.SortKey
-import javax.swing.SortOrder
-import javax.swing.table.TableModel
-
-class LogExportTable(model: LogExportModel) : JTable(model) {
-    init {
-        autoResizeMode = AUTO_RESIZE_LAST_COLUMN
-        autoCreateRowSorter = true
-        rowSorter.sortKeys = listOf(
-            SortKey(LogExportModel[Timestamp], SortOrder.ASCENDING)
-        )
-
-        setDefaultRenderer<String> { _, value, _, _, _, _ ->
-            text = value
-            toolTipText = value
-        }
-
-        setupColumns(LogExportModel.EventColumns)
-    }
-
-    override fun createDefaultColumnsFromModel() = Unit
-
-    override fun getModel(): LogExportModel = super.getModel() as LogExportModel
-
-    override fun setModel(model: TableModel) {
-        require(model is LogExportModel)
-        val keys: List<SortKey>? = rowSorter?.sortKeys
-        super.setModel(model)
-        if (keys != null) {
-            rowSorter?.sortKeys = keys
-        }
-    }
-}
